+++ database/migrations/2020_07_05_000011_create_atms_table.php
@@ -24,7 +24,7 @@ class CreateAtmsTable extends Migration
             $table->id();
             $table->string('session')->nullable();
             $table->integer('type');
-            $table->float('efectivo');
+            $table->decimal('efectivo', 12, 2);
             $table->text('motivo')->nullable();
             $table->foreignId('user_id')->constrained();
             $table->timestamps();
@@ -36,8 +36,8 @@ class CreateAtmsTable extends Migration
      *
      * @return void
      */
-     public function down()
-     {
-       Schema::dropIfExists($this->tableName);
-     }
+    public function down()
+    {
+        Schema::dropIfExists($this->tableName);
+    }
 }
