+++ database/migrations/2020_09_21_000013_create_sales_items_table.php
@@ -23,8 +23,8 @@ class CreateSalesItemsTable extends Migration
         Schema::create($this->tableName, function (Blueprint $table) {
             $table->id();
             $table->integer('cant');
-            $table->float('price');
-            $table->float('subtotal');
+            $table->decimal('price', 12, 2);
+            $table->decimal('subtotal', 12, 2);
             $table->integer('inStorage')->default(0);
             $table->integer('out')->nullable()->default(0);
             $table->string('session', 100);
@@ -40,8 +40,8 @@ class CreateSalesItemsTable extends Migration
      *
      * @return void
      */
-     public function down()
-     {
-       Schema::dropIfExists($this->tableName);
-     }
+    public function down()
+    {
+        Schema::dropIfExists($this->tableName);
+    }
 }
